+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## HEAD
+
+* Make `posix-spawn` optional (@janko-m)
+
 ## 4.4.0 (2018-04-12)
 
 * Add `:method` option to `Down::Http` for specifying the request method (@janko-m)